Break recursive processing when directory is empty.
The `curl` command appears to return blank when it lists an empty
directory. This, in turn, sets $_FILES to empty.

The `while` loop over `$_FILES` then yields an empty `$_LINE` and the
rest of the function proceeds, eventually appending a `/` to the end of
the next URL to test -- but this new URL is effectively the same as the one
just listed!

This leads to infinite recursion checking the same empty directory over
and over again, e.g. list `foo/bar/`, then `foo/bar//`, then
`foo/bar///` etc etc.

This commit adds a test for the base case -- the directory is empty --
and breaks appropriately.